start of the offline profile
the offline profile is supposed to be a helper module that can be used
to switch agents to using a local repo / pre stage files, etc. It uses
run stages from stdlib to ensure that repos are setup first.

el6_repo.pp points el6 hosts to the masters local repo