Update firewall profile to work with centos 6.5
This is really more for systems that don't have a default firewall at
all. In CentOS 6.5 it was observed that if the firewall service
(iptables) tried to start before any rules had been defined, startup
would fail. Changing the dependencies of the default pre firewall rules
to come before the service makes this work again under those
circumstances.