Add an example directory and first example class
As the profile module is intended as a sampler, it makes sense to have
an example directory with interesting use cases demonstrated, even if
they are either self-contained or outside the scope of the more
integrated class set included in the module.

The profile::example::nfsserver class was originally created for a
puppetlabs/firewall module webinar.