Update the local repo definition with OS version
Previously the local repo was a single-directory repository. This worked
while all we had using it was RedHat 6 based systems, but when we added
RedHat 7 to the mix we realized that we had to have separate yum repos
for different platforms. This commit replaces the single "rpms"
directory with an OS and version sensitive directory name, such as
"RedHat-6". We are intentionally not adding architecture.