Add simple Windows registry example
The new/old profile::example::registry class is taken from code in the
https://github.com/puppetlabs/saleseng-demo-environment/tree/master/modules/profile
repository. This was an old version of the seteam/profile module, prior
to it being split into a separate repository.

The registry example can be used to demonstrate idempotency on Windows
systems by creating registry keys if they don't exist, and enforcing
their state and content.